feat(loglevels): Made default loglevel INFO
Made the log package accept an integer for loglevel instead of strings. This was done so that a user could write a commandline program with a verbose flag such as -v or -vv in typer. This flag gets interpreted as 1 or 2 and so the log packages needs to be made for numbers.