refactor: Added custom loglevel and switch to OOP design
Added a custom log level to the logger called VERBOSE. This falls in between INFO and DEBUG. This level will allow normal runtime information to be displayed but also for verbose info to be displayed if desired. This change was accomplished by Wisperwind's solution (https://stackoverflow.com/questions/2183233/how-to-add-a-custom-loglevel-to-pythons-logging-facility/35804945). Now the create_logger method and get_logger methods are staticmethods

BREAKING CHANGE: