Feature: jtb324 - 8/20/25 2:18pm
files updated - pyproject.toml,src/log/logger.py

Previously the logging module required you to write to a file and then you could optionally write to an output stream. There was no way to turn of the filehandler. This behavior is no longer ideal because the testing moddule of DRIVE doesn't need to write to a file. Updated the code so that by default it will still write to a file but the user can also disable this and then the logging module will automatically write to STDOUT instead.