Refactor: jtb324 - 8/20/20 - 2:29pm
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
       new file:   pdm.lock
       modified:   pyproject.toml
       modified:   src/log/logger.py

----
Updated the type hint for filename to make sure it was clear that this could be None now and made sure that branching was correct and there wouldn't be any bugs caused by trying to create a new Path object with output_path if it was provided. Also added black for project formatting