Refactor: - jtb324 - 8/21/25 - 7:38am
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
       modified:   pyproject.toml
       modified:   src/log/logger.py

Fixed a string formatting issue on lines 53-59 for the logger:
----

Template string multiplication was not being performed right since the string multipplication was not in an f-string. Fixed this so now the border should show up. This change includes an updated patch number