Feature(sampling): - jtb324 - 9/26/25 - 10:17am: --- On branch sampling Changes to be committed: 	modified:   pyproject.toml 	new file:   src/log/sampling_filters/__init__.py 	new file:   src/log/sampling_filters/filter_interface.py 	new file:   src/log/sampling_filters/random_sampling_filter.py
began adding sampling procedures to the logger
----
DRIVE has also produced very large logging files when the loglevel is
set to low (specifically debug but to some degree the verbose setting too).
I added a filter interface that allows us to specify a filter object
that implements the filter method to sample log records. Currently this
interface is only implemented in a RandomPropFilter class. This class
randomly selects which records to keep based on a proportion provided at
the object's initialization. This behavior can be customized to only
apply to records below or at a specific loglevel if the user so desires