Refactor(logger.py): - jtb324 - 9/26/25 - 5:42pm --- On branch sampling Changes to be committed: 	modified:   src/log/__init__.py 	modified:   src/log/logger.py
refactored configure method of CustomLogger
---
The code was starting to require a lot of parameters for the
configuration of the logger. The parameters dealt with different aspects
of the logger such as how to configure the handler, how to configure
verbosity and what to print to, and how to configure the formatter. I
abstracting this away to make *Opts classes for the Handler, Logger, and
Formatter. This change makes the code more like a builder pattern and
also allows for each Opts object to validate their respective values.
Hopefully we did not abstract too early