Feature(logger.py): - jtb324 - 9/27/25 - 8:11am --- On branch sampling Changes to be committed: 	modified:   src/log/logger.py
added ability to use RotatingFileHandler
---
One additional feature that can decrease the ultimate size of the log
file is to have the logger automatically start a new log file after it
reaches a certain size. This can easy be accomplished by using the
logging.handlers.RotatingFileHandler. By default it creates a new file
every 1024 bytes and only saves 5 backups but this can be customized by
the user