Feature(json_formatter.py): - jtb324 - 9/28/29 - 9:10pm --- On branch sampling Changes to be committed: 	modified:   src/log/formatters/__init__.py 	modified:   src/log/formatters/json_formatter.py 	modified:   src/log/logger.py
created custom formatter for writing JSON
---
Sometimes having structured logging in the output log file is more
beneficial than having a formatted string. This module contains a custom
logger for JSON. The json object records a lot of the runtime
information and can include additional context if it is passed through a
LoggerAdapter. I added a parameter to the FormatterOpts class called
"write_json" that would allow the user to switch between the
JsonFormatter and the normal logging.FileFormattter.