Feature(logger.py): - jtb324 - 9/29/25 - 11:11am --- On branch refactor Changes to be committed: 	modified:   src/log/logger.py
allow users to add context to the logger:
---
Created a bind method that binds context to the parent logger using the
logging.LoggerAdapter library. This method accepts the context as a
dictionary and returns the ModifiedLoggerAdapter which functions as a
new logger with the appropriate context bound to it. If the user
provides new context within the log statements (such as in log.info or
log.debug), then the user defined values will override the context in
the adapter.